refactor(precompile): create crypto submodule for secp256r1
Move `verify_signature` and its feature-gated implementations
(p256 vs aws-lc-rs) into a dedicated `crypto.rs` submodule,
separating pure cryptographic logic from the precompile runner.

Closes part of #2782
