+++ crates/precompile/src/secp256r1/mod.rs
@@ -6,9 +6,13 @@
 //! The main purpose of this precompile is to verify ECDSA signatures that use the secp256r1, or
 //! P256 elliptic curve. The [`P256VERIFY`] const represents the implementation of this precompile,
 //! with the address that it is currently deployed at.
+pub mod crypto;
+
+pub(crate) use crypto::verify_signature;
+
 use crate::{
-    crypto, u64_to_address, Precompile, PrecompileError, PrecompileId, PrecompileOutput,
-    PrecompileResult,
+    crypto as crypto_provider, u64_to_address, Precompile, PrecompileError, PrecompileId,
+    PrecompileOutput, PrecompileResult,
 };
 use primitives::{alloy_primitives::B512, Bytes, B256};
 
@@ -92,41 +96,7 @@ pub fn verify_impl(input: &[u8]) -> bool {
     // x, y: public key
     let pk = <&B512>::try_from(&input[96..160]).unwrap();
 
-    crypto().secp256r1_verify_signature(&msg.0, &sig.0, &pk.0)
-}
-
-pub(crate) fn verify_signature(msg: &[u8; 32], sig: &[u8; 64], pk: &[u8; 64]) -> Option<()> {
-    cfg_if::cfg_if! {
-        if #[cfg(feature = "p256-aws-lc-rs")] {
-            use aws_lc_rs::{digest, signature::{self, UnparsedPublicKey}};
-
-            // Construct a Digest from the raw prehashed message bytes.
-            let digest = digest::Digest::import_less_safe(msg, &digest::SHA256).ok()?;
-
-            // Build uncompressed public key: 0x04 || x || y
-            let mut pubkey_bytes = [0u8; 65];
-            pubkey_bytes[0] = 0x04;
-            pubkey_bytes[1..].copy_from_slice(pk);
-
-            let public_key = UnparsedPublicKey::new(&signature::ECDSA_P256_SHA256_FIXED, &pubkey_bytes);
-
-            public_key.verify_digest(&digest, sig).ok()
-        } else {
-            use p256::{
-                ecdsa::{signature::hazmat::PrehashVerifier, Signature, VerifyingKey},
-                EncodedPoint,
-            };
-
-            // Can fail only if the input is not exact length.
-            let signature = Signature::from_slice(sig).ok()?;
-            // Decode the public key bytes (x,y coordinates) using EncodedPoint
-            let encoded_point = EncodedPoint::from_untagged_bytes(&(*pk).into());
-            // Create VerifyingKey from the encoded point
-            let public_key = VerifyingKey::from_encoded_point(&encoded_point).ok()?;
-
-            public_key.verify_prehash(msg, &signature).ok()
-        }
-    }
+    crypto_provider().secp256r1_verify_signature(&msg.0, &sig.0, &pk.0)
 }
 
 #[cfg(test)]
